Handle files over 2GB (#1065)
* Handle files over 2GB

* Use Guava's checkArgument

+++ gcloud-java-contrib/gcloud-java-nio/src/main/java/com/google/cloud/storage/contrib/nio/CloudStorageReadChannel.java
@@ -118,7 +118,7 @@ final class CloudStorageReadChannel implements SeekableByteChannel {
       if (newPosition == position) {
         return this;
       }
-      channel.seek((int) newPosition);
+      channel.seek(newPosition);
       position = newPosition;
       return this;
     }
